requirements.txt
ipython
python-dotenv
ollama
pillow
torch
diffusers
